Clean up semantic metric view debt
